requirements.txt
anyio==2.2.0
asks==2.4.12
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
decorator==4.0.11
h11==0.12.0
idna==3.2
imageio==2.1.2
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
lxml==4.6.3
mega.py==1.0.8
moviepy==0.2.3.2
numpy==1.21.2
outcome==1.1.0
pathlib==1.0.1
Pillow==8.3.2
pycryptodome==3.10.4
requests==2.26.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
tenacity==5.1.5
tqdm==4.11.2
trio==0.19.0
urllib3==1.26.7
